requirements.txt
gradio
httpx
feedparser
beautifulsoup4
pyalex
python-dotenv
python-dateutil